% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{getSerieDataFrame}
\alias{getSerieDataFrame}
\title{Get a data.frame from an series Vector}
\usage{
getSerieDataFrame(series, idSerie)
}
\arguments{
\item{series}{A vector containing data series. This is the vector returned by
\code{\link{getSeriesData}}.}

\item{idSerie}{A string intentifying the series required, it can only be one.}
}
\value{
A data.frame containing the required data series.
}
\description{
This is an utility function, it allows to obtain a data.frame from the vector returned by
\code{\link{getSeriesData}}.
}
\examples{

\dontrun{
## You need a valid token to run the example
setToken("token")

series <- getSeriesMetadata(c("SF43718","SF46410","SF46407"))

serie <- getSerieDataFrame(series, "SF43718")
}

}
